pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ghost-writer2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "jsonschema>=4.25.1",
    "openai>=1.102.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "streamlit>=1.49.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/ghost_writer"]